Cargo clippy fixes

+++ src/loops/discovery.rs
@@ -151,7 +151,7 @@ pub fn start_discovery_loop(state: &InternalState) {
                 .with_option("max_execution_time", "240");
 
             let range_end_sec = now_secs() as u32;
-            let range_start_sec = range_end_sec - local_config.look_back_hours as u32 * 3600;
+            let range_start_sec = range_end_sec - local_config.look_back_hours * 3600;
             let start = Instant::now();
 
             // ----------------------------------------------------------------
